social_media_api/requirements - Copy.txt
altgraph==0.17.5
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
chardet==4.0.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
django-taggit==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
feedparser==6.0.12
h11==0.16.0
idna==2.10
outcome==1.3.0.post0
packaging==25.0
pefile==2024.8.26
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
PyJWT==2.11.0
PySocks==1.7.1
pywin32-ctypes==0.2.3
requests==2.25.1
selenium==4.38.0
setuptools==80.9.0
sgmllib3k==1.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.9.0
wsproto==1.2.0
